+++ src/FluentlyHttpClient/FluentHttpRequest.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Net.Http;
+using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
@@ -65,6 +66,11 @@ namespace FluentlyHttpClient
 		/// Gets or sets a key/value collection that can be used to share data within the scope of request/response.
 		/// </summary>
 		public IDictionary<object, object> Items { get; protected set; }
+
+		/// <summary>
+		/// Formatters to be used for content negotiation for "Accept" and also sending formats. e.g. (JSON, XML)
+		/// </summary>
+		public MediaTypeFormatterCollection Formatters { get; set; }
 
 		/// <summary>
 		/// Initializes a new instance.