+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 ### Features
 
 - **request builder:** implement `WithItem` which allows to set custom items that can be used to share data within the scope of request, response, and middleware.
+- **request builder:** add `WithUserAgent` extension
 
 - **http client builder:** formatter JSON is now configured with camelcase property names by default.
 
 - **logger middleware:** add extension method `UseLogging`.
+
+- **consts:** add constants for headers and auth schemes `HeaderTypes` and `AuthSchemeTypes`
 
 
  ## [1.0.0](https://github.com/sketch7/FluentlyHttpClient/compare/0.3.0...1.0.0) (2017-06-30)