+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - **request builder:** implement `WithItem` which allows to set custom items that can be used to share data within the scope of request, response, and middleware.
 - **request builder:** add `WithUserAgent` extension
 
+- **header builder:** extract interface `IFluentHttpHeaderBuilder` which both `FluentHttpClientBuilder` and `FluentHttpRequestBuilder` implements.
+This is done to be able to share implementations for extensions methods across `FluentHttpClientBuilder` and `FluentHttpRequestBuilder`
+
 - **http client builder:** formatter JSON is now configured with camelcase property names by default.
 
 - **logger middleware:** add extension method `UseLogging`.