+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Features
 
 - **request builder:** implement `WithItem` which allows to set custom items that can be used to share data within the scope of request, response, and middleware.
-- **request builder:** add `WithUserAgent` extension
+- **request builder:** add `WithUserAgent` extension.
+- **request builder:** add validation for request when 'GET' and has body content, to not be allowed. As it will blow up the underlying HttpClient.
 
 - **header builder:** extract interface `IFluentHttpHeaderBuilder`, which both `FluentHttpClientBuilder` and `FluentHttpRequestBuilder` implements.
 This is done to be able to share implementations for extensions methods across `FluentHttpClientBuilder` and `FluentHttpRequestBuilder`