+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Fluently Http Changelog
 
- [*vNext*](https://github.com/sketch7/FluentlyHttpClient/compare/0.1.0...1.1.0) (201X-X-X)
+ [*vNext*](https://github.com/sketch7/FluentlyHttpClient/compare/1.1.0...1.2.0) (201X-X-X)
+
+## [1.1.0](https://github.com/sketch7/FluentlyHttpClient/compare/1.0.0...1.1.0) (2017-07-02)
 
 ### Features